wip: move hasher modules to utils
