+++ src/bn254/fq12.rs
@@ -862,7 +862,7 @@ mod test {
     use crate::bn254::fq::{fq_to_nibbles, nibbles_to_fq, Fq};
     use crate::bn254::fq12::Fq12;
     use crate::bn254::fq6::Fq6;
-    use crate::bn254::utils::{fq12_push, fq12_push_not_montgomery, fq2_push, fq2_push_not_montgomery, fq6_push_not_montgomery, read_script_from_file};
+    use crate::bn254::utils::{fq12_push, fq12_push_not_montgomery, fq2_push, fq2_push_not_montgomery, fq6_push_not_montgomery, hash_fp12, hash_fp12_192, hash_fp12_with_hints, hash_fp6, read_script_from_file};
     use crate::pseudo::NMUL;
     use crate::{execute_script_without_stack_limit, treepp::*};
     use ark_ff::AdditiveGroup;
@@ -1273,204 +1273,6 @@ mod test {
             run(script);
         }
     }
-
-
-
-    fn hash_fp6() -> Script {
-
-        let hash_64b_75k = read_script_from_file("/Users/manishbista/Documents/alpenlabs/BitVM/blake3_bin/blake3_64b_75k.bin");
-        let hash_128b_168k = read_script_from_file("/Users/manishbista/Documents/alpenlabs/BitVM/blake3_bin/blake3_128b_168k.bin");
-
-        script!{
-            for _ in 0..=4 {
-                {Fq::toaltstack()}
-            }
-            { fq_to_nibbles() }
-            { Fq::fromaltstack() }
-            { fq_to_nibbles() }
-            {hash_64b_75k.clone()}
-            { nibbles_to_fq() }
-
-            { Fq::fromaltstack() }
-            {fq_to_nibbles()}
-            { Fq::fromaltstack() }
-            {fq_to_nibbles()}
-            { Fq::fromaltstack() }
-            {fq_to_nibbles()}
-            { Fq::fromaltstack() }
-            {fq_to_nibbles()}
-            { hash_128b_168k.clone() }
-
-            for i in 1..=1 {
-                for _ in 0..9 {
-                    {64*i + 8} OP_ROLL
-                }
-                {fq_to_nibbles()}
-            }
-            { hash_64b_75k }
-            {nibbles_to_fq()}
-        } 
-    }
-
-
-
-    fn hash_fp12() -> Script {
-
-        let hash_64b_75k = read_script_from_file("/Users/manishbista/Documents/alpenlabs/BitVM/blake3_bin/blake3_64b_75k.bin");
-        let hash_128b_168k = read_script_from_file("/Users/manishbista/Documents/alpenlabs/BitVM/blake3_bin/blake3_128b_168k.bin");
-
-        script!{
-            for i in 0..=10 {
-                {Fq::toaltstack()}
-            }
-
-            // first part
-            { fq_to_nibbles() }
-            { Fq::fromaltstack() }
-            { fq_to_nibbles() }
-            {hash_64b_75k.clone()}
-            { nibbles_to_fq() }
-
-            { Fq::fromaltstack() }
-            {fq_to_nibbles()}
-            { Fq::fromaltstack() }
-            {fq_to_nibbles()}
-            { Fq::fromaltstack() }
-            {fq_to_nibbles()}
-            { Fq::fromaltstack() }
-            {fq_to_nibbles()}
-            { hash_128b_168k.clone() }
-
-
-            for i in 1..=1 {
-                for _ in 0..9 {
-                    {64*i + 8} OP_ROLL
-                }
-                {fq_to_nibbles()}
-            }
-            {hash_64b_75k.clone()}
-            {nibbles_to_fq()}
-
-            // second part
-
-            { Fq::fromaltstack() }
-            {fq_to_nibbles()}
-            { Fq::fromaltstack() }
-            {fq_to_nibbles()}
-            {hash_64b_75k.clone()}
-            { nibbles_to_fq() }
-            
-
-            { Fq::fromaltstack() }
-            {fq_to_nibbles()}
-            { Fq::fromaltstack() }
-            {fq_to_nibbles()}
-            { Fq::fromaltstack() }
-            {fq_to_nibbles()}
-            { Fq::fromaltstack() }
-            {fq_to_nibbles()}
-            { hash_128b_168k.clone() }
-
-            for i in 1..=1 {
-                for _ in 0..9 {
-                    {64*i + 8} OP_ROLL
-                }
-                {fq_to_nibbles()}
-            }
-            {hash_64b_75k.clone()}
-
-            // wrap up
-            for i in 1..=1 {
-                for _ in 0..9 {
-                    {64*i + 8} OP_ROLL
-                }
-                {fq_to_nibbles()}
-            }
-            {hash_64b_75k.clone()}
-            {nibbles_to_fq()}
-
-        } 
-    }
-
-    fn hash_fp12_192() -> Script {
-        let hash_64b_75k = read_script_from_file("/Users/manishbista/Documents/alpenlabs/BitVM/blake3_bin/blake3_64b_75k.bin");
-        let hash_192b_252k = read_script_from_file("/Users/manishbista/Documents/alpenlabs/BitVM/blake3_bin/blake3_192b_252k.bin");
-        script! {
-            for i in 0..=10 {
-                {Fq::toaltstack()}
-            }
-            {fq_to_nibbles() }
-            for i in 0..5 {
-                { Fq::fromaltstack()}
-                {fq_to_nibbles()}
-            }
-            {hash_192b_252k.clone()}
-            {nibbles_to_fq()}
-
-            for i in 0..6 {
-                { Fq::fromaltstack()}
-                {fq_to_nibbles()}
-            }
-            {hash_192b_252k}
-            for i in 0..9 {
-                {64+8} OP_ROLL
-            }
-            { fq_to_nibbles() }
-            {hash_64b_75k}
-            {nibbles_to_fq()}
-        }
-    }
-
-    // 6Fp_hash
-    // fp6
-    fn hash_fp12_with_hints() -> Script {
-
-        let hash_64b_75k = read_script_from_file("/Users/manishbista/Documents/alpenlabs/BitVM/blake3_bin/blake3_64b_75k.bin");
-        let hash_128b_168k = read_script_from_file("/Users/manishbista/Documents/alpenlabs/BitVM/blake3_bin/blake3_128b_168k.bin");
-
-        script!{
-            for i in 0..=4 {
-                {Fq::toaltstack()}
-            }
-
-            { fq_to_nibbles() }
-            { Fq::fromaltstack() }
-            { fq_to_nibbles() }
-            {hash_64b_75k.clone()}
-            { nibbles_to_fq() }
-
-            { Fq::fromaltstack() }
-            {fq_to_nibbles()}
-            { Fq::fromaltstack() }
-            {fq_to_nibbles()}
-            { Fq::fromaltstack() }
-            {fq_to_nibbles()}
-            { Fq::fromaltstack() }
-            {fq_to_nibbles()}
-            { hash_128b_168k.clone() }
-
-
-            for i in 1..=1 {
-                for _ in 0..9 {
-                    {64*i + 8} OP_ROLL
-                }
-                {fq_to_nibbles()}
-            }
-            {hash_64b_75k.clone()}
-
-            // wrap up
-            for i in 1..=1 {
-                for _ in 0..9 {
-                    {64*i + 8} OP_ROLL
-                }
-                {fq_to_nibbles()}
-            }
-            {hash_64b_75k.clone()}
-            {nibbles_to_fq()}
-
-        } 
-    }
-
 
 
     #[test]
