+++ src/bn254/utils.rs
@@ -1,6 +1,8 @@
 // utils for push fields into stack
 use crate::bn254::ell_coeffs::EllCoeff;
 use crate::bn254::ell_coeffs::G2Prepared;
+use crate::bn254::fq::fq_to_nibbles;
+use crate::bn254::fq::nibbles_to_fq;
 use crate::bn254::fr::Fr;
 use crate::bn254::{fq12::Fq12, fq2::Fq2};
 use ark_ec::{bn::BnConfig, AffineRepr};
@@ -1382,6 +1384,203 @@ pub fn read_script_from_file(file_path: &str) -> Script {
 
 
 
+pub fn hash_fp6() -> Script {
+
+    let hash_64b_75k = read_script_from_file("/Users/manishbista/Documents/alpenlabs/BitVM/blake3_bin/blake3_64b_75k.bin");
+    let hash_128b_168k = read_script_from_file("/Users/manishbista/Documents/alpenlabs/BitVM/blake3_bin/blake3_128b_168k.bin");
+
+    script!{
+        for _ in 0..=4 {
+            {Fq::toaltstack()}
+        }
+        { fq_to_nibbles() }
+        { Fq::fromaltstack() }
+        { fq_to_nibbles() }
+        {hash_64b_75k.clone()}
+        { nibbles_to_fq() }
+
+        { Fq::fromaltstack() }
+        {fq_to_nibbles()}
+        { Fq::fromaltstack() }
+        {fq_to_nibbles()}
+        { Fq::fromaltstack() }
+        {fq_to_nibbles()}
+        { Fq::fromaltstack() }
+        {fq_to_nibbles()}
+        { hash_128b_168k.clone() }
+
+        for i in 1..=1 {
+            for _ in 0..9 {
+                {64*i + 8} OP_ROLL
+            }
+            {fq_to_nibbles()}
+        }
+        { hash_64b_75k }
+        {nibbles_to_fq()}
+    } 
+}
+
+
+
+pub fn hash_fp12() -> Script {
+
+    let hash_64b_75k = read_script_from_file("/Users/manishbista/Documents/alpenlabs/BitVM/blake3_bin/blake3_64b_75k.bin");
+    let hash_128b_168k = read_script_from_file("/Users/manishbista/Documents/alpenlabs/BitVM/blake3_bin/blake3_128b_168k.bin");
+
+    script!{
+        for i in 0..=10 {
+            {Fq::toaltstack()}
+        }
+
+        // first part
+        { fq_to_nibbles() }
+        { Fq::fromaltstack() }
+        { fq_to_nibbles() }
+        {hash_64b_75k.clone()}
+        { nibbles_to_fq() }
+
+        { Fq::fromaltstack() }
+        {fq_to_nibbles()}
+        { Fq::fromaltstack() }
+        {fq_to_nibbles()}
+        { Fq::fromaltstack() }
+        {fq_to_nibbles()}
+        { Fq::fromaltstack() }
+        {fq_to_nibbles()}
+        { hash_128b_168k.clone() }
+
+
+        for i in 1..=1 {
+            for _ in 0..9 {
+                {64*i + 8} OP_ROLL
+            }
+            {fq_to_nibbles()}
+        }
+        {hash_64b_75k.clone()}
+        {nibbles_to_fq()}
+
+        // second part
+
+        { Fq::fromaltstack() }
+        {fq_to_nibbles()}
+        { Fq::fromaltstack() }
+        {fq_to_nibbles()}
+        {hash_64b_75k.clone()}
+        { nibbles_to_fq() }
+        
+
+        { Fq::fromaltstack() }
+        {fq_to_nibbles()}
+        { Fq::fromaltstack() }
+        {fq_to_nibbles()}
+        { Fq::fromaltstack() }
+        {fq_to_nibbles()}
+        { Fq::fromaltstack() }
+        {fq_to_nibbles()}
+        { hash_128b_168k.clone() }
+
+        for i in 1..=1 {
+            for _ in 0..9 {
+                {64*i + 8} OP_ROLL
+            }
+            {fq_to_nibbles()}
+        }
+        {hash_64b_75k.clone()}
+
+        // wrap up
+        for i in 1..=1 {
+            for _ in 0..9 {
+                {64*i + 8} OP_ROLL
+            }
+            {fq_to_nibbles()}
+        }
+        {hash_64b_75k.clone()}
+        {nibbles_to_fq()}
+
+    } 
+}
+
+pub fn hash_fp12_192() -> Script {
+    let hash_64b_75k = read_script_from_file("/Users/manishbista/Documents/alpenlabs/BitVM/blake3_bin/blake3_64b_75k.bin");
+    let hash_192b_252k = read_script_from_file("/Users/manishbista/Documents/alpenlabs/BitVM/blake3_bin/blake3_192b_252k.bin");
+    script! {
+        for i in 0..=10 {
+            {Fq::toaltstack()}
+        }
+        {fq_to_nibbles() }
+        for i in 0..5 {
+            { Fq::fromaltstack()}
+            {fq_to_nibbles()}
+        }
+        {hash_192b_252k.clone()}
+        {nibbles_to_fq()}
+
+        for i in 0..6 {
+            { Fq::fromaltstack()}
+            {fq_to_nibbles()}
+        }
+        {hash_192b_252k}
+        for i in 0..9 {
+            {64+8} OP_ROLL
+        }
+        { fq_to_nibbles() }
+        {hash_64b_75k}
+        {nibbles_to_fq()}
+    }
+}
+
+// 6Fp_hash
+// fp6
+pub fn hash_fp12_with_hints() -> Script {
+
+    let hash_64b_75k = read_script_from_file("/Users/manishbista/Documents/alpenlabs/BitVM/blake3_bin/blake3_64b_75k.bin");
+    let hash_128b_168k = read_script_from_file("/Users/manishbista/Documents/alpenlabs/BitVM/blake3_bin/blake3_128b_168k.bin");
+
+    script!{
+        for i in 0..=4 {
+            {Fq::toaltstack()}
+        }
+
+        { fq_to_nibbles() }
+        { Fq::fromaltstack() }
+        { fq_to_nibbles() }
+        {hash_64b_75k.clone()}
+        { nibbles_to_fq() }
+
+        { Fq::fromaltstack() }
+        {fq_to_nibbles()}
+        { Fq::fromaltstack() }
+        {fq_to_nibbles()}
+        { Fq::fromaltstack() }
+        {fq_to_nibbles()}
+        { Fq::fromaltstack() }
+        {fq_to_nibbles()}
+        { hash_128b_168k.clone() }
+
+
+        for i in 1..=1 {
+            for _ in 0..9 {
+                {64*i + 8} OP_ROLL
+            }
+            {fq_to_nibbles()}
+        }
+        {hash_64b_75k.clone()}
+
+        // wrap up
+        for i in 1..=1 {
+            for _ in 0..9 {
+                {64*i + 8} OP_ROLL
+            }
+            {fq_to_nibbles()}
+        }
+        {hash_64b_75k.clone()}
+        {nibbles_to_fq()}
+
+    } 
+}
+
+
+
 #[cfg(test)]
 mod test {
     use super::*;
@@ -1667,10 +1866,14 @@ mod test {
         f1.mul_by_034(&ark_bn254::Fq2::ONE, &c1new, &c2new);
         let (hinted_script, hints) = Fq12::hinted_mul_by_34(f, c1new, c2new);
 
+        let hash_128b_168k = read_script_from_file("/Users/manishbista/Documents/alpenlabs/BitVM/blake3_bin/blake3_128b_168k.bin");
+
+
         let script = script! {
             for hint in hints { 
                 { hint.push() }
             }
+
             { fq12_push_not_montgomery(f) }
             { fq2_push_not_montgomery(c1new) }
             { fq2_push_not_montgomery(c2new) }
@@ -1678,11 +1881,35 @@ mod test {
             { fq2_push_not_montgomery(c1new) }
             { fq2_push_not_montgomery(c2new) }
             { hinted_script.clone() }
-            { fq12_push_not_montgomery(f1) }
-            { Fq12::equalverify() }
-            {Fq2::drop()}
-            {Fq2::drop()}
-            {Fq12::drop()}
+
+            for _ in 0..12 { // save claim
+                {Fq::toaltstack()}
+            }
+
+            // hash c1new and c2new
+            {Fq::roll(1)}
+            {Fq::roll(2)}
+            {Fq::roll(3)}
+            { Fq::toaltstack() }
+            { Fq::toaltstack() }
+            { Fq::toaltstack() }
+            {fq_to_nibbles()} // 0
+            { Fq::fromaltstack()}
+            {fq_to_nibbles()}
+            { Fq::fromaltstack()}
+            {fq_to_nibbles()}
+            { Fq::fromaltstack()}
+            {fq_to_nibbles()}
+            {hash_128b_168k.clone()}
+            {nibbles_to_fq()}
+
+            {hash_fp12() }
+
+            for _ in 0..12 { // save claim
+                {Fq::fromaltstack()}
+            }
+            { hash_fp12_192() }
+
             OP_TRUE
         };
         let len = script.len();
